almost fixed completely  made it use updateQ main problem is that i see that it teleports when not supposed to

+++ Bounties/src/sim/app/bounties/Task.java
@@ -91,7 +91,7 @@ public class Task implements Real, Fixed2D{
         lastFinishedRobotID =  robotid;
         finishedTime = timestamp;
     }
-    public long getLastFinishedRobotID(){
+    public int getLastFinishedRobotID(){
         return lastFinishedRobotID;
     }
     public long getLastFinishedTime() {
